Migrate InputParameter to lembas.param module
